refactor: Rename functions and files
